+++ packages/gamut-kit/CHANGELOG.md
@@ -1,7 +1,23 @@
-# Change Log
+## 0.6.592 (2026-03-26)
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+### 🩹 Fixes
+
+- minor type changes due to typescript upgrade ([#3254](https://github.com/Codecademy/gamut/pull/3254))
+
+### 🧱 Updated Dependencies
+
+- Updated gamut-illustrations to 0.58.9
+- Updated gamut-patterns to 0.10.28
+- Updated gamut-styles to 17.13.1
+- Updated gamut-icons to 9.57.2
+- Updated gamut-tests to 5.3.3
+- Updated variance to 0.26.1
+- Updated gamut to 68.2.2
+
+### ❤️ Thank You
+
+- Copilot @Copilot
+- jakemhiller @jakemhiller
 
 ### [0.6.591](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-kit@0.6.590...@codecademy/gamut-kit@0.6.591) (2026-03-25)