+++ packages/gamut-tests/CHANGELOG.md
@@ -1,7 +1,17 @@
-# Change Log
+## 5.3.3 (2026-03-26)
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+### 🩹 Fixes
+
+- minor type changes due to typescript upgrade ([#3254](https://github.com/Codecademy/gamut/pull/3254))
+
+### 🧱 Updated Dependencies
+
+- Updated gamut-styles to 17.13.1
+
+### ❤️ Thank You
+
+- Copilot @Copilot
+- jakemhiller @jakemhiller
 
 ### [5.3.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-tests@5.3.1...@codecademy/gamut-tests@5.3.2) (2026-03-25)